fix(protocol): keep validator generation out of install
